// block-performance-1.0.0/src/library/aws-lightsail/instances-create/type.ts
export interface Input {
    credentials: { accessKeyId: string; secretAccessKey: string };
    region?: string;
    instanceNames: string[];
    availabilityZone: string;
    blueprintId: string;
    bundleId: string;
    userData?: string;
    keyPairName?: string;
    tags?: { key: string; value: string }[];
    ipAddressType?: "dualstack" | "ipv4";
}

export interface OutputOk {
    operations: {
        createdAt: string;
        id: string;
        isTerminal: boolean;
        location: { availabilityZone: string; regionName: string };
        operationType: string;
        resourceName: string;
        resourceType: string;
        status: string;
        statusChangedAt: string;
    }[];
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export const copy_output = copy_output_ok;